check verify

+++ Project/demo/src/main/java/com/fptu/swp391/se1839/oemevwarrantymanagement/dto/response/DashboardClaimSummaryResponse.java
@@ -14,5 +14,5 @@ import lombok.experimental.FieldDefaults;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class DashboardClaimSummaryResponse {
     Long count;
-    Long emegency;
+    int emegency;
 }
